+++ core/ui/src/main/java/com/contraomnese/weather/core/ui/widgets/AirQualityItem.kt
@@ -73,7 +73,7 @@ fun AirQualityItem(
             modifier = Modifier
                 .fillMaxWidth()
                 .height(itemHeight6),
-            current = index.toFloat()
+            current = index
         )
         Row(
             modifier = Modifier.fillMaxWidth(),
