+++ core/ui/src/main/java/com/contraomnese/weather/core/ui/widgets/UvIndexItem.kt
@@ -69,7 +69,7 @@ fun UvIndexItem(
             modifier = Modifier
                 .fillMaxWidth()
                 .height(itemHeight8),
-            current = index.toFloat(),
+            current = index,
         )
 
         Text(
